requirements.txt
flask>=3.0,<4
werkzeug>=3.0,<4
psycopg2-binary>=2.9.10,<3
python-dotenv>=1.0,<2
httpx>=0.27,<1
